bug: totalAward를 구할때 long을 사용하도록 한다

+++ src/main/java/lotto/game/Round.java
@@ -36,7 +36,7 @@ public class Round {
 
     public long totalAward(List<Integer> winNumbers) {
         return getResult(winNumbers).entrySet().stream()
-                .mapToInt(entry -> entry.getKey().award * entry.getValue())
+                .mapToLong(entry -> (long) entry.getKey().award * entry.getValue())
                 .sum();
     }
 }
